requirements.txt
streamlit
fastapi
uvicorn
pydantic
twilio
python-dotenv
pandas
numpy
requests
openai
httpx

# Optional: for working with notebooks
jupyter
ipykernel
